apps/backend/requirements.txt
flask>=2.0.0
flask-cors>=4.0.0
pymysql>=1.0.0
reportlab>=4.0.0
pyjwt>=2.0.0
